requirements.txt
alembic==1.11.2
blinker==1.6.2
click==8.1.6
colorama==0.4.6
distlib==0.3.7
dominate==2.8.0
filelock==3.12.2
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysqlclient==2.2.0
platformdirs==3.10.0
PyMySQL==1.1.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
virtualenv==20.24.2
visitor==0.1.3
Werkzeug==2.3.6
WTForms==3.0.1
